Customer_support/Code/document_pipeline/requirements.txt
python-dotenv
openai
qdrant-client
docling
requests
aiohttp
